# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: 邻里车位 API
  description: 阶段 2 免费车位共享 MVP 接口
  version: 0.2.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
paths:
  /api/v1/vehicles/{id}:
    put:
      tags:
      - vehicle-controller
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Vehicle'
    delete:
      tags:
      - vehicle-controller
      operationId: deactivate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
...
  /api/v1/vehicles:
    get:
      tags:
      - vehicle-controller
      operationId: list
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'
    post:
      tags:
      - vehicle-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleRequest'
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Vehicle'
...
components:
  schemas:
    VehicleRequest:
      required:
      - plateNumber
      - vehicleType
      type: object
      properties:
        plateNumber:
          maxLength: 24
          minLength: 0
          type: string
        vehicleType:
          maxLength: 32
          minLength: 0
          type: string
    Vehicle:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        plateNumber:
          type: string
        vehicleType:
          type: string
        verificationStatus:
          type: string
          enum:
          - PENDING
          - APPROVED
          - REJECTED
        active:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT